Dev/GCalendar/requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
aiohttp==3.8.5
asyncio==3.4.3